Add animation for background gradient
Calculate required animation for the gradient background based on
the current/next page. There is sometimes a small flash when the
next page enters due to the order React renders the components in.
This can be seen by console logging each render from the Layout
component. May need to try and obtain more detailed control of
the rendering to fix it.